Implement Object Constructor for books
Remove the id attribute on the remove button and
add id attribute to the li element. When the user
clicks "Remove", find the corresponding item of
that remove button in the bookList and then use
its removeBook() method which has been declared
in its prototype.

Use the currentId as an alternative to the
bookList index so that when the user remove an
item (not the last item) from the array, the next
id won't repeat the previous one.